Car/requirements.txt
Django==3.1.1
Gunicorn==20.0.4
django-heroku==0.3.1
djangorestframework==3.12.1
requests==2.24.0